Changes
Added a new FastQuestFill that works properly and uses hashmaps

Reworked packages. Since we cannot use properties in our parallel use case, we will have 2 dictionaries in AgentFunctions.h. They will pair a package to an alias.

Get---Package will be called inside the spell, and grab the correct package for the corresponding alias. This will then be used in AddPackageOverride